requirements.txt
boolean.py==4.0
dash[cache]==2.15.0
dash_bootstrap_components==1.5.0
diskcache==5.6.3
paperscraper==0.2.10
Pyarrow==15.0.0
